update rewards file and CRRA utilities
