+++ HARK/ConsumptionSaving/tests/test_ConsMedModel.py
@@ -1,5 +1,3 @@
-from HARK.utilities import CRRAutility_inv
-import numpy as np
 from HARK.ConsumptionSaving.ConsMedModel import MedShockConsumerType
 import unittest
 
@@ -7,7 +5,7 @@ import unittest
 class testMedShockConsumerType(unittest.TestCase):
     def setUp(self):
 
-        self.agent = MedicalExample = MedShockConsumerType()
+        self.agent = MedShockConsumerType()
 
     def test_solution(self):
