+++ HARK/utilities.py
@@ -43,6 +43,8 @@ def memoize(obj):
 # ==============================================================================
 # ============== Some basic function tools  ====================================
 # ==============================================================================
+
+
 def get_arg_names(function):
     """
     Returns a list of strings naming all of the arguments for the passed function.
@@ -108,430 +110,6 @@ class NullFunc(object):
             return 10000.0
 
 
-# ==============================================================================
-# ============== Define utility functions        ===============================
-# ==============================================================================
-def CRRAutility(c, gam):
-    """
-    Evaluates constant relative risk aversion (CRRA) utility of consumption c
-    given risk aversion parameter gam.
-
-    Parameters
-    ----------
-    c : float
-        Consumption value
-    gam : float
-        Risk aversion
-
-    Returns
-    -------
-    (unnamed) : float
-        Utility
-
-    Tests
-    -----
-    Test a value which should pass:
-    >>> c, gamma = 1.0, 2.0    # Set two values at once with Python syntax
-    >>> utility(c=c, gam=gamma)
-    -1.0
-    """
-
-    if gam == 1:
-        return np.log(c)
-    else:
-        return c ** (1.0 - gam) / (1.0 - gam)
-
-
-def uFunc_CRRA_stone_geary(c, CRRA, stone_geary):
-    """
-    Evaluates Stone-Geary version of a constant relative risk aversion (CRRA)
-    utility of consumption c wiht given risk aversion parameter CRRA and
-    Stone-Geary intercept parameter stone_geary
-
-    Parameters
-    ----------
-    c : float
-        Consumption value
-    CRRA : float
-        Relative risk aversion
-    stone_geary : float
-        Intercept in Stone-Geary utility
-    Returns
-    -------
-    (unnamed) : float
-        Utility
-
-    Tests
-    -----
-    Test a value which should pass:
-    >>> c, CRRA, stone_geary = 1.0, 2.0, 0.0
-    >>> utility(c=c, CRRA=CRRA, stone_geary=stone_geary )
-    -1.0
-    """
-    if CRRA == 1:
-        return np.log(stone_geary + c)
-    else:
-        return (stone_geary + c) ** (1.0 - CRRA) / (1.0 - CRRA)
-
-
-def uPFunc_CRRA_stone_geary(c, CRRA, stone_geary):
-    """
-    Marginal utility of Stone-Geary version of a constant relative risk aversion (CRRA)
-    utility of consumption c wiht given risk aversion parameter CRRA and
-    Stone-Geary intercept parameter stone_geary
-
-    Parameters
-    ----------
-    c : float
-        Consumption value
-    CRRA : float
-        Relative risk aversion
-    stone_geary : float
-        Intercept in Stone-Geary utility
-    Returns
-    -------
-    (unnamed) : float
-        marginal utility
-
-    """
-    return (stone_geary + c) ** (-CRRA)
-
-
-def uPPFunc_CRRA_stone_geary(c, CRRA, stone_geary):
-    """
-    Marginal marginal utility of Stone-Geary version of a CRRA utilty function
-    with risk aversion parameter CRRA and Stone-Geary intercept parameter stone_geary
-
-    Parameters
-    ----------
-    c : float
-        Consumption value
-    CRRA : float
-        Relative risk aversion
-    stone_geary : float
-        Intercept in Stone-Geary utility
-    Returns
-    -------
-    (unnamed) : float
-        marginal utility
-
-    """
-    return (-CRRA) * (stone_geary + c) ** (-CRRA - 1)
-
-
-def CRRAutilityP(c, gam):
-    """
-    Evaluates constant relative risk aversion (CRRA) marginal utility of consumption
-    c given risk aversion parameter gam.
-
-    Parameters
-    ----------
-    c : float
-        Consumption value
-    gam : float
-        Risk aversion
-
-    Returns
-    -------
-    (unnamed) : float
-        Marginal utility
-    """
-
-    if gam == 1:
-        return 1 / c
-
-    return c**-gam
-
-
-def CRRAutilityPP(c, gam):
-    """
-    Evaluates constant relative risk aversion (CRRA) marginal marginal utility of
-    consumption c given risk aversion parameter gam.
-
-    Parameters
-    ----------
-    c : float
-        Consumption value
-    gam : float
-        Risk aversion
-
-    Returns
-    -------
-    (unnamed) : float
-        Marginal marginal utility
-    """
-
-    return -gam * c ** (-gam - 1.0)
-
-
-def CRRAutilityPPP(c, gam):
-    """
-    Evaluates constant relative risk aversion (CRRA) marginal marginal marginal
-    utility of consumption c given risk aversion parameter gam.
-
-    Parameters
-    ----------
-    c : float
-        Consumption value
-    gam : float
-        Risk aversion
-
-    Returns
-    -------
-    (unnamed) : float
-        Marginal marginal marginal utility
-    """
-
-    return (gam + 1.0) * gam * c ** (-gam - 2.0)
-
-
-def CRRAutilityPPPP(c, gam):
-    """
-    Evaluates constant relative risk aversion (CRRA) marginal marginal marginal
-    marginal utility of consumption c given risk aversion parameter gam.
-
-    Parameters
-    ----------
-    c : float
-        Consumption value
-    gam : float
-        Risk aversion
-
-    Returns
-    -------
-    (unnamed) : float
-        Marginal marginal marginal marginal utility
-    """
-
-    return -(gam + 2.0) * (gam + 1.0) * gam * c ** (-gam - 3.0)
-
-
-def CRRAutility_inv(u, gam):
-    """
-    Evaluates the inverse of the CRRA utility function (with risk aversion para-
-    meter gam) at a given utility level u.
-
-    Parameters
-    ----------
-    u : float
-        Utility value
-    gam : float
-        Risk aversion
-
-    Returns
-    -------
-    (unnamed) : float
-        Consumption corresponding to given utility value
-    """
-    if gam == 1:
-        return np.exp(u)
-    else:
-        return ((1.0 - gam) * u) ** (1 / (1.0 - gam))
-
-
-def CRRAutilityP_inv(uP, gam):
-    """
-    Evaluates the inverse of the CRRA marginal utility function (with risk aversion
-    parameter gam) at a given marginal utility level uP.
-
-    Parameters
-    ----------
-    uP : float
-        Marginal utility value
-    gam : float
-        Risk aversion
-
-    Returns
-    -------
-    (unnamed) : float
-        Consumption corresponding to given marginal utility value.
-    """
-    return uP ** (-1.0 / gam)
-
-
-def CRRAutility_invP(u, gam):
-    """
-    Evaluates the derivative of the inverse of the CRRA utility function (with
-    risk aversion parameter gam) at a given utility level u.
-
-    Parameters
-    ----------
-    u : float
-        Utility value
-    gam : float
-        Risk aversion
-
-    Returns
-    -------
-    (unnamed) : float
-        Marginal consumption corresponding to given utility value
-    """
-    if gam == 1:
-        return np.exp(u)
-    else:
-        return ((1.0 - gam) * u) ** (gam / (1.0 - gam))
-
-
-def CRRAutilityP_invP(uP, gam):
-    """
-    Evaluates the derivative of the inverse of the CRRA marginal utility function
-    (with risk aversion parameter gam) at a given marginal utility level uP.
-
-    Parameters
-    ----------
-    uP : float
-        Marginal utility value
-    gam : float
-        Risk aversion
-
-    Returns
-    -------
-    (unnamed) : float
-        Consumption corresponding to given marginal utility value
-    """
-    return (-1.0 / gam) * uP ** (-1.0 / gam - 1.0)
-
-
-def CARAutility(c, alpha):
-    """
-    Evaluates constant absolute risk aversion (CARA) utility of consumption c
-    given risk aversion parameter alpha.
-
-    Parameters
-    ----------
-    c: float
-        Consumption value
-    alpha: float
-        Risk aversion
-
-    Returns
-    -------
-    (unnamed): float
-        Utility
-    """
-    return 1 - np.exp(-alpha * c) / alpha
-
-
-def CARAutilityP(c, alpha):
-    """
-    Evaluates constant absolute risk aversion (CARA) marginal utility of
-    consumption c given risk aversion parameter alpha.
-
-    Parameters
-    ----------
-    c: float
-        Consumption value
-    alpha: float
-        Risk aversion
-
-    Returns
-    -------
-    (unnamed): float
-        Marginal utility
-    """
-    return np.exp(-alpha * c)
-
-
-def CARAutilityPP(c, alpha):
-    """
-    Evaluates constant absolute risk aversion (CARA) marginal marginal utility
-    of consumption c given risk aversion parameter alpha.
-
-    Parameters
-    ----------
-    c: float
-        Consumption value
-    alpha: float
-        Risk aversion
-
-    Returns
-    -------
-    (unnamed): float
-        Marginal marginal utility
-    """
-    return -alpha * np.exp(-alpha * c)
-
-
-def CARAutilityPPP(c, alpha):
-    """
-    Evaluates constant absolute risk aversion (CARA) marginal marginal marginal
-    utility of consumption c given risk aversion parameter alpha.
-
-    Parameters
-    ----------
-    c: float
-        Consumption value
-    alpha: float
-        Risk aversion
-
-    Returns
-    -------
-    (unnamed): float
-        Marginal marginal marginal utility
-    """
-    return alpha**2.0 * np.exp(-alpha * c)
-
-
-def CARAutility_inv(u, alpha):
-    """
-    Evaluates inverse of constant absolute risk aversion (CARA) utility function
-    at utility level u given risk aversion parameter alpha.
-
-    Parameters
-    ----------
-    u: float
-        Utility value
-    alpha: float
-        Risk aversion
-
-    Returns
-    -------
-    (unnamed): float
-        Consumption value corresponding to u
-    """
-    return -1.0 / alpha * np.log(alpha * (1 - u))
-
-
-def CARAutilityP_inv(u, alpha):
-    """
-    Evaluates the inverse of constant absolute risk aversion (CARA) marginal
-    utility function at marginal utility uP given risk aversion parameter alpha.
-
-    Parameters
-    ----------
-    u: float
-        Utility value
-    alpha: float
-        Risk aversion
-
-    Returns
-    -------
-    (unnamed): float
-        Consumption value corresponding to uP
-    """
-    return -1.0 / alpha * np.log(u)
-
-
-def CARAutility_invP(u, alpha):
-    """
-    Evaluates the derivative of inverse of constant absolute risk aversion (CARA)
-    utility function at utility level u given risk aversion parameter alpha.
-
-    Parameters
-    ----------
-    u: float
-        Utility value
-    alpha: float
-        Risk aversion
-
-    Returns
-    -------
-    (unnamed): float
-        Marginal onsumption value corresponding to u
-    """
-    return 1.0 / (alpha * (1.0 - u))
-
-
 # =======================================================
 # ================ Other useful functions ===============
 # =======================================================
@@ -593,146 +171,11 @@ def construct_assets_grid(parameters):
     return aXtraGrid
 
 
-def cobb_douglas(x, alpha, factor):
-    """
-    Evaluates Cobb Douglas utility at quatitites of goods consumed `x`
-    given elasticity parameters `alpha` and efficiency parameter `factor`.
-
-    Parameters
-    ----------
-    x : np.ndarray
-        Quantities of goods consumed. Last axis must index goods.
-    alpha : np.ndarray
-        Elasticity parameters for each good. Must be consistent with `x`.
-    factor : float
-        Multiplicative efficiency parameter. (e.g. TFP in production function)
-
-    Returns
-    -------
-    (unnamed) : np.ndarray
-        Utility
-
-    """
-
-    return factor * np.sum(x**alpha, axis=-1)
-
-
-def cobb_douglas_p(x, alpha, factor, arg=0):
-    """
-    Evaluates the marginal utility of consumption indexed by `arg` good at
-    quantities of goods consumed `x` given elasticity parameters `alpha`
-    and efficiency parameter `factor`.
-
-    Parameters
-    ----------
-    x : np.ndarray
-        Quantities of goods consumed. Last axis must index goods.
-    alpha : np.ndarray
-        Elasticity parameters for each good. Must be consistent with `x`.
-    factor : float
-        Multiplicative efficiency parameter.
-    arg : int
-        Index of good to evaluate marginal utility.
-
-    Returns
-    -------
-    (unnamed) : np.ndarray
-        Utility
-    """
-
-    return cobb_douglas(x, alpha, factor) * alpha[arg] / x[..., arg]
-
-
-def cobb_douglas_pp(x, alpha, factor, args=(0, 1)):
-    """
-    Evaluates the marginal marginal utility of consumption indexed by `args`
-    at quantities of goods consumed `x` given elasticity parameters `alpha`
-    and efficiency parameter `factor`.
-
-    Parameters
-    ----------
-    x : np.ndarray
-        Quantities of goods consumed. Last axis must index goods.
-    alpha : np.ndarray
-        Elasticity parameters for each good. Must be consistent with `x`.
-    factor : float
-        Multiplicative efficiency parameter.
-    args : tuple(int)
-        Indexes of goods to evaluate marginal utility. `args[0]` is the
-        index of the first derivative taken, and `args[1]` is the index of
-        the second derivative taken.
-
-    Returns
-    -------
-    (unnamed) : np.ndarray
-        Utility
-    """
-
-    if args[0] == args[1]:
-        coeff = alpha[args[1]] - 1
-    else:
-        coeff = alpha[args[1]]
-
-    return cobb_douglas_p(x, alpha, factor, args[0]) * coeff / x[..., args[1]]
-
-
-def cobb_douglas_pn(x, alpha, factor, args=()):
-    """
-    Evaluates the nth marginal utility of consumption indexed by `args`
-    at quantities of goods consumed `x` given elasticity parameters `alpha`
-    and efficiency parameter `factor`.
-
-    Parameters
-    ----------
-    x : np.ndarray
-        Quantities of goods consumed. Last axis must index goods.
-    alpha : np.ndarray
-        Elasticity parameters for each good. Must be consistent with `x`.
-    factor : float
-        Multiplicative efficiency parameter.
-    args : tuple(int)
-        Indexes of goods to evaluate marginal utility. `args[0]` is the
-        index of the first derivative taken, and `args[1]` is the index of
-        the second derivative taken. This function works by recursively taking
-        derivatives, so `args` can be of any length.
-
-    Returns
-    -------
-    (unnamed) : np.ndarray
-        Utility
-    """
-
-    if isinstance(args, int):
-        args = (args,)
-
-    if len(args):
-        counts = dict(zip(*np.unique(args, return_counts=True)))
-        idx = args[-1]
-        coeff = alpha[idx] - counts[idx] + 1
-        new_args = tuple(list(args)[:-1])
-        return cobb_douglas_pn(x, alpha, factor, new_args) * coeff / x[..., idx]
-    else:
-        return cobb_douglas(x, alpha, factor)
-
-
-def const_elast_subs(x, alpha, subs, factor, power):
-
-    return factor * np.sum(alpha * x**subs, axis=-1) ** (power / subs)
-
-
-def const_elast_subs_p(x, alpha, subs, factor, power, arg=0):
-
-    return (
-        const_elast_subs(x, alpha, factor * power / subs, subs, power - subs)
-        * alpha[arg]
-        * subs
-        * x[..., arg] ** (subs - 1)
-    )
-
-
 # ==============================================================================
 # ============== Functions for generating state space grids  ===================
 # ==============================================================================
+
+
 def make_grid_exp_mult(ming, maxg, ng, timestonest=20):
     """
     Make a multi-exponentially spaced grid.
@@ -780,6 +223,8 @@ def make_grid_exp_mult(ming, maxg, ng, timestonest=20):
 # ==============================================================================
 # ============== Uncategorized general functions  ===================
 # ==============================================================================
+
+
 def calc_weighted_avg(data, weights):
     """
     Generates a weighted average of simulated data.  The Nth row of data is averaged
